Marvell: Add library class for NonDiscoverable device init routines
To abstract the initialization required for non-discoverable devices, which
is often platform specific (i.e., enabling VBUS gpios for USB), introduce
a NonDiscoverableInitLib for use by the NonDiscoverable code, for which each
platform can supply its own version.

Add dummy version of the library for A70x0 platform in order to satisfy
compilation. It will be extended in future with proper VBUS control,
as well as with the support for other boards.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Marcin Wojtas <mw@semihalf.com>