Marvell/Library: ComPhyLib: Set preset request vector to preset4
This preset is used during the initial part of the equalization
master phase. According to hardware measurements, our PCIe hardware
can use preset4 or preset6. However, when we try both preset4 and
preset6, for some GEN3 devices, the link might be established as GEN1
and not GEN3.
Probably, this happens due to a timeout (when 2 presets are used).

This patch updates the MAC and the PHY to use only preset4, in such
case no link issues are observed.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>